lab02/requirements.txt
PyQt5==5.15.4
transformers==4.24.0
torch>=1.0.0
python-dotenv==0.19.2
